SHOPS/RealTime/requirements.txt
# Real-time Parser MCP Server
httpx>=0.25.0
psycopg2-binary>=2.9.0
beautifulsoup4>=4.12.0
mcp>=0.9.0
